simplify the config and template generation
