+++ trainite/config/base.py
@@ -1,12 +1,65 @@
-# __COLLATE_IMPORT__
-from typing import Self
+from typing import Any, Self, Literal
+from pydantic import BaseModel, ConfigDict, Field, model_validator
 
-from pydantic import (
-    BaseModel,
-    ConfigDict,
-    Field,
-    model_validator,
-)
+
+class ModelConfig(BaseModel):
+    model_config = ConfigDict(extra="allow", validate_assignment=True)
+    target: str = Field(alias="_target_")
+
+    @model_validator(mode="before")
+    @classmethod
+    def validate_from_any_model(cls, data: Any) -> Any:
+        if isinstance(data, BaseModel):
+            return data.model_dump(by_alias=True)
+        return data
+
+
+class PreprocessorConfig(BaseModel):
+    model_config = ConfigDict(extra="allow", validate_assignment=True)
+    target: str = Field(alias="_target_")
+
+    @model_validator(mode="before")
+    @classmethod
+    def validate_from_any_model(cls, data: Any) -> Any:
+        if isinstance(data, BaseModel):
+            return data.model_dump(by_alias=True)
+        return data
+
+
+class DatasetConfig(BaseModel):
+    model_config = ConfigDict(extra="allow", validate_assignment=True)
+    target: str = Field(alias="_target_")
+
+    @model_validator(mode="before")
+    @classmethod
+    def validate_from_any_model(cls, data: Any) -> Any:
+        if isinstance(data, BaseModel):
+            return data.model_dump(by_alias=True)
+        return data
+
+
+class TransformConfig(BaseModel):
+    model_config = ConfigDict(extra="allow", validate_assignment=True)
+    target: str = Field(alias="_target_")
+
+    @model_validator(mode="before")
+    @classmethod
+    def validate_from_any_model(cls, data: Any) -> Any:
+        if isinstance(data, BaseModel):
+            return data.model_dump(by_alias=True)
+        return data
+
+
+class CollateFnConfig(BaseModel):
+    model_config = ConfigDict(extra="allow", validate_assignment=True)
+    target: str = Field(alias="_target_")
+
+    @model_validator(mode="before")
+    @classmethod
+    def validate_from_any_model(cls, data: Any) -> Any:
+        if isinstance(data, BaseModel):
+            return data.model_dump(by_alias=True)
+        return data
 
 
 class OutputConfig(BaseModel):
@@ -20,19 +73,26 @@ class OptimizerConfig(BaseModel):
     target: str = Field(alias="_target_", default="torch.optim.AdamW")
     lr: float = Field(default=1e-3, gt=0.0)
 
+    @model_validator(mode="before")
+    @classmethod
+    def validate_from_any_model(cls, data: Any) -> Any:
+        if isinstance(data, BaseModel):
+            return data.model_dump(by_alias=True)
+        return data
+
 
 class DataLoaderConfig(BaseModel):
     model_config = ConfigDict(extra="allow", validate_assignment=True)
     batch_size: int = Field(default=32, gt=0)
     shuffle: bool = False
     num_workers: int = Field(default=2, ge=0)
-    collate_fn: BaseModel | None = None
+    collate_fn: CollateFnConfig | None = None
 
 
 class SplitConfig(BaseModel):
     model_config = ConfigDict(validate_assignment=True)
-    dataset: BaseModel
-    transform: BaseModel | None = None
+    dataset: DatasetConfig
+    transform: TransformConfig | None = None
     dataloader: DataLoaderConfig = Field(default_factory=DataLoaderConfig)
 
 
@@ -46,8 +106,8 @@ class DataConfigBase(BaseModel):
 class DataWithAutoSplit(BaseModel):
     model_config = ConfigDict(validate_assignment=True)
     # Option 2: Automatic splitting
-    dataset: BaseModel
-    transform: BaseModel | None = None
+    dataset: DatasetConfig
+    transform: TransformConfig | None = None
     dataloader: DataLoaderConfig = Field(default_factory=DataLoaderConfig)
     test_ratio: float = Field(default=0.2, ge=0.0, lt=0.5)
     val_ratio: float = Field(default=0.1, gt=0.0, lt=0.3)
@@ -60,3 +120,35 @@ class DataWithAutoSplit(BaseModel):
         if train_ratio > 0.9:
             raise ValueError("train ratio (1.0 - test_ratio - val_ratio) must be less than or equal to 0.9")
         return self
+
+
+class TrainerConfig(BaseModel):
+    model_config = ConfigDict(extra="allow", validate_assignment=True)
+    epochs: int = Field(default=3, gt=0)
+    log_every_steps: int = Field(default=10, gt=0)
+    early_stopping_patience: int | None = Field(default=3, gt=0)
+    # Inference logging
+    inference_every_epochs: int | None = Field(default=None, gt=0)
+    inference_num_samples: int = Field(default=5, gt=0)
+    max_inference_new_tokens: int = Field(default=16, gt=0)
+    grad_clip_norm: float | None = Field(default=None, gt=0.0)
+
+    @model_validator(mode="before")
+    @classmethod
+    def validate_from_any_model(cls, data: Any) -> Any:
+        if isinstance(data, BaseModel):
+            return data.model_dump(by_alias=True)
+        return data
+
+
+class ProjectConfig(BaseModel):
+    model_config = ConfigDict(validate_assignment=True)
+    preprocessor: PreprocessorConfig | None = None
+    model: ModelConfig
+    optimizer: OptimizerConfig = Field(default_factory=OptimizerConfig)
+    data: DataConfigBase | DataWithAutoSplit
+    trainer: TrainerConfig = Field(default_factory=TrainerConfig)
+    output: OutputConfig
+    logger: Literal["tensorboard", "wandb"] = "tensorboard"
+    seed: int = 42
+    device: str | None = None
