+++ trainite/config/trainers.py
@@ -1,0 +1,5 @@
+from trainite.config.base import TrainerConfig as BaseTrainerConfig
+
+
+class TrainerConfig(BaseTrainerConfig):
+    pass
