+++ trainite/preprocessors/__init__.py
@@ -1,3 +1,4 @@
-from trainite.preprocessors.char_tokenizer import UNIVERSAL_VOCAB, CharTokenizer, CharTokenizerConfig
+from trainite.preprocessors.char_tokenizer import UNIVERSAL_VOCAB, CharTokenizer
+from trainite.config.preprocessors import CharTokenizerConfig
 
 __all__ = ["CharTokenizer", "CharTokenizerConfig", "UNIVERSAL_VOCAB"]
