+++ trainite/preprocessors/char_tokenizer.py
@@ -2,7 +2,7 @@ import string
 from typing import Any
 
 import torch
-from pydantic import BaseModel, ConfigDict, Field
+
 
 # Hardcoded universal vocabulary: all printable ASCII characters
 UNIVERSAL_VOCAB = string.ascii_letters + string.digits + string.punctuation + " "
@@ -147,11 +147,3 @@ class CharTokenizer:
             return tensor_out
 
         return out
-
-
-class CharTokenizerConfig(BaseModel):
-    model_config = ConfigDict(extra="allow", validate_assignment=True)
-    target: str = Field(
-        default="trainite.preprocessors.char_tokenizer.CharTokenizer",
-        alias="_target_",
-    )
